lab8/requirements.txt
azure-ai-textanalytics==5.3.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
matplotlib>=3.7.2
nltk>=3.8.1
gensim>=4.3.1 